Change way to show/hide infowindow
add show-property and show-value to info window to decide which scope
value to use to show/hide the infowindow.
if the scope value of show-property is equal to show-value then the info
window is shown otherwise hidden
eg.
show-property = 'name'
show-value = 'value'
the info window is shown when $scope.name == 'value'